Desafio v7 - Implementação e melhoria na autorização, verificar bug no require

+++ NodeJs - Bootcamp/Desafio-final/repository/livro.repository.js
@@ -1,5 +1,6 @@
 const db = require("../model/livraria.models.js");
-
+const infoRepository = require("./info.livro.repository.js");
+const vendaService = require("../service/venda.service.js");
 async function criarLivro(livro) {
   try {
     return await db.livros.create(livro);
@@ -11,7 +12,7 @@ async function criarLivro(livro) {
 async function atualizarLivro(livro) {
   try {
     await db.livros.update(livro, { where: { livroId: livro.livroId } });
-    return await buscarLivro(livro.livroId  );
+    return await buscarLivro(livro.livroId);
   } catch (err) {
     throw err;
   }
@@ -19,7 +20,21 @@ async function atualizarLivro(livro) {
 
 async function buscarLivro(id) {
   try {
-    return await db.livros.findByPk(id, { raw: true });
+    const livro = await db.livros.findByPk(id, { raw: true });
+    const informacoes = await infoRepository.buscarInfo(id);
+    const retorno = Object.assign(livro, informacoes);
+    return retorno;
+  } catch (err) {
+    throw err;
+  }
+}
+
+async function buscarLivro2(id) {
+  try {
+    const livro = await db.livros.findByPk(id, { raw: true });
+    const informacoes = await infoRepository.buscarInfo(id);
+    const retorno = Object.assign(livro, informacoes);
+    return retorno;
   } catch (err) {
     throw err;
   }
@@ -42,6 +57,10 @@ async function buscarLivroPorAutor(id) {
 
 async function deletarLivro(id) {
   try {
+    const vendas = await vendaService.buscarVendaByLivro(id);
+    if (vendas) {
+      throw new Error("Não foi possivel deletar - Há vendas deste livro ");
+    }
     await db.livros.destroy({ where: { livroId: id } });
   } catch (err) {
     throw err;
@@ -55,4 +74,5 @@ module.exports = {
   deletarLivro,
   buscarTodosLivros,
   buscarLivroPorAutor,
+  buscarLivro2
 };
